Fixed a bug on RemoveStashLimit
By a report on the nexus page, when the player have more than 50 arrows, both bows broke and its not possible to use it anymore. This happens cuz when the bowController tries to access InventoryView, and it return a number greater than the array itself. So by now I did a lazy fix, I just limit the index used to abtain the arrow view, so instead of usin something like: Mathf.Max(LocalPlayer.Inventory.AmountOf(_ammoItemId, false) - 1, 0) I have used a limit to the view, of 50 arrows, this will result in a minor bug on the numbers of arrow, but only the text. I will fix this eventually.